message sending, read description
Files changed: Conversation, CSVReadWrite, and Message

Messages: changed fields and toString method so it could work the way I needed it to in conversations

CSVReadWrite: added 3 methods replaceLine and append each have a new function that does not automatically do the indexing automatically. Other method is getNextIndex it just return a string of the next index that will be used.

Conversation: made it so that conversations can happen and made a way to send messages and display all messages in a conversation, I included a main there that shows off some of the parts of it, makes sure to download the messages and conversations csvs that are included for it to work. If you run the test take note of the csv files before and after you run it and make sure they update.