conversations implemented into application
conversations now work, user is shown conversations if they have none they are prompted to make one. User can select conversation and then messages from it will be shown. Then user is given choice of go back to conversation list, send message to the conversation it is in, quit, or make a new conversation. I have uploaded it and do not have any lingering bugs that I know of.